feat: move PSP to coordinates module
