+++ src/galax/coordinates/__init__.py
@@ -1,0 +1,11 @@
+""":mod:`galax.coordinates`."""
+
+from . import _base, _core, _utils
+from ._base import *
+from ._core import *
+from ._utils import *
+
+__all__: list[str] = []
+__all__ += _base.__all__
+__all__ += _core.__all__
+__all__ += _utils.__all__
