+++ tests/smoke/dynamics/test_package.py
@@ -1,15 +1,9 @@
 """Testing :mod:`galax.dynamics` module."""
 
 import galax.dynamics as gd
-from galax.dynamics._dynamics import base, core, mockstream, orbit
+from galax.dynamics._dynamics import mockstream, orbit
 
 
 def test_all() -> None:
     """Test the `galax.potential` API."""
-    assert set(gd.__all__) == {
-        "mockstream",
-        *base.__all__,
-        *core.__all__,
-        *orbit.__all__,
-        *mockstream.__all__,
-    }
+    assert set(gd.__all__) == {"mockstream", *orbit.__all__, *mockstream.__all__}
